rename getInFund to createCashFlow
