+++ src/features/CreateCashFlow/view/components/ConfigurationCommitment/ConfigurationCommitment.style.ts
